Revert changes made due to threading (atomic primitives and concurrent lists)
Reverting all changes made because of atomic primitives and concurrent lists seems to fix most of the bugs which are currently in the refactor-dev branch. This could be a temporary version to deploy to production. I like the use of atomicity and concurrency but we have to do those changes step by step to see when and why bugs occurr.